Moved viewmodel factory tests into new directory
