+++ app/src/test/java/com/soen490chrysalis/papilio/viewModel/factories/UserProfileViewModelFactoryTest.kt
@@ -1,7 +1,9 @@
-package com.soen490chrysalis.papilio.viewModel
+package com.soen490chrysalis.papilio.viewModel.factories
 
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import com.google.firebase.auth.FirebaseAuth
+import com.soen490chrysalis.papilio.viewModel.UserProfileViewModel
+import com.soen490chrysalis.papilio.viewModel.UserProfileViewModelFactory
 import io.mockk.every
 import io.mockk.mockkStatic
 import org.junit.Rule
